adjust DDR3 static tuning parameters for king credie build
Adjusted hard-coded parameters for memories that don't support
write levelling calibrations for the latest build.

Probably in the future, we'll need to have a factory cal procedure
that determines the optimal tuning for the two most common forms
of DDR3 memory, and commits them to I2C, so we don't have to
manually adjust this every time we have a new board batch.

Impedances and relative flight times shift based upon the vagaries
of individual board manufacturers, hence the variability in the cal
data.